getOrderData test case refactored
